python/requirements.txt
uvicorn==0.27.1
typing==3.7.4.3
fastapi==0.110.0
matplot==0.1.9
pandas==2.2.1
scikit-learn==1.4.1.post1
seaborn==0.13.2
pymongo==4.6.2
pydantic-settings==2.2.1
mysql-connector-python==8.3.0

